put server code in a separate subpackage
