+++ puzzle/server/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .factory import create_app
