+++ puzzle/server/blueprints/public/views.py
