+++ puzzle/server/blueprints/variants/templates/macros/gt_calls.html
