+++ puzzle/server/blueprints/variants/templates/sv_variant.html
