+++ puzzle/server/blueprints/variants/templates/variant.html
