+++ puzzle/server/blueprints/variants/views.py
