+++ puzzle/server/factory.py
