+++ puzzle/server/settings.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 import os.path
 
+from puzzle.plugins import VcfPlugin
 from .blueprints import public_bp, variants_bp
-from .plugins import VcfPlugin
 
 PROJECT_NAME = __name__.split('.')[0]
