+++ puzzle/server/static/styles.css
